pkb/requirements.gateway.txt
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
pyyaml>=6.0
python-telegram-bot>=21.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-api>=1.24.0
